requirements.txt
pandas==1.5.3
arrow
astropy
datetime
geopy
pytz
streamlit==1.26.0
influxdb_client
reportlab
Pillow
plotly
matplotlib
dataframe_image
matplotlib